Alumno Terminado
Está todo Terminado
lo que falta es Asignatura
"Cursar, Aprobar o Perder regularidad de una asignatura para un alumno
Para este caso, lo que ocurre es que se intentará modificar el estado de una de las
asignaturas. Asumiremos cuando se crea el alumno que se creará una Lista de
Asignaturas asociadas a él y la misma puede estar hardcodeada en el código.
PUT: /alumno/{idAlumno}/asignatura/{idAsignatura}"

falta hacer eso solamente (Con manejo de errores).
El manejo de errores de todo lo demás esta implementado... (Por ahí se pueda mejorar más, pero iré viendo apenas vaya viendo el código).

+++ src/main/java/ar/edu/utn/frbb/tup/business/AsignaturaService.java
@@ -3,7 +3,7 @@ package ar.edu.utn.frbb.tup.business;
 import ar.edu.utn.frbb.tup.model.Asignatura;
 
 public interface AsignaturaService {
-    Asignatura getAsignatura(int materiaId, long dni);
+    Asignatura getAsignatura(int materiaId, int dni);
 
     void actualizarAsignatura(Asignatura a);
 }
