+++ src/main/java/ar/edu/utn/frbb/tup/model/Alumno.java
@@ -8,17 +8,17 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Alumno {
-    private long id;
+    private Integer id;
 
     private String nombre;
     private String apellido;
-    private long dni;
+    private int dni;
 
     private List<Asignatura> asignaturas;
 
     public Alumno() {
     }
-    public Alumno(String nombre, String apellido, long dni) {
+    public Alumno(String nombre, String apellido, int dni) {
         this.nombre = nombre;
         this.apellido = apellido;
         this.dni = dni;
@@ -35,7 +35,7 @@ public class Alumno {
         this.apellido = apellido;
     }
 
-    public void setDni(long dni) {
+    public void setDni(int dni) {
         this.dni = dni;
     }
 
@@ -47,7 +47,7 @@ public class Alumno {
         return apellido;
     }
 
-    public long getDni() {
+    public int getDni() {
         return dni;
     }
 
@@ -105,11 +105,11 @@ public class Alumno {
 
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 }
