+++ src/main/java/ar/edu/utn/frbb/tup/model/dto/AlumnoDto.java
@@ -4,7 +4,7 @@ public class AlumnoDto {
 
     String nombre;
     String apellido;
-    long dni;
+    int dni;
 
     public String getNombre() {
         return nombre;
@@ -22,11 +22,11 @@ public class AlumnoDto {
         this.apellido = apellido;
     }
 
-    public long getDni() {
+    public int getDni() {
         return dni;
     }
 
-    public void setDni(long dni) {
+    public void setDni(int dni) {
         this.dni = dni;
     }
 }
